refactor: move helper check functions to compatibility_helper file
