+++ formats/json/testdata/enum/curr_addition.json
@@ -1,0 +1,16 @@
+{
+    "title": "summaries",
+    "properties": {
+        "roles": {
+            "title": "Organization roles",
+            "type": "string",
+            "enum": [
+                "producer",
+                "licensor",
+                "processor",
+                "host",
+                "super_user"
+            ]
+        }
+    }
+}
