+++ formats/json/testdata/enum/curr_removal.json
@@ -1,0 +1,13 @@
+{
+    "title": "summaries",
+    "properties": {
+        "roles": {
+            "title": "Organization roles",
+            "type": "string",
+            "enum": [
+                "producer",
+                "licensor"
+            ]
+        }
+    }
+}
