+++ formats/json/testdata/enum/non_enum.json
@@ -1,0 +1,8 @@
+{
+    "title": "summaries",
+    "properties": {
+        "roles": {
+            "title": "Organization roles"
+        }
+    }
+}
